pyproject.toml
[tool.poetry]
name = "doba-michaels-update"
version = "0.1.0"
description = ""
authors = ["Sam Benning <sam.b@elitecommercegroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.2.1"
requests = "^2.32.3"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
